perf: add overwrite protection to do_customize()

+++ src/databased/dbmanager.py
@@ -2,7 +2,6 @@ import argshell
 from pathier import Pathier
 
 import databased
-
 from databased import dbparsers
 
 
@@ -151,12 +150,15 @@ class DBManager(argshell.ArgShell):
         Expects one argument: the name of the custom dbmanager.
         This will be used to name the generated file as well as several components in the file content."""
         custom_file = (Pathier.cwd() / command).with_suffix(".py")
-        variable_name = "_".join(word for word in command.lower().split())
-        class_name = "".join(word.capitalize() for word in command.split())
-        content = (Pathier(__file__).parent / "custom_manager.py").read_text()
-        content = content.replace("CustomManager", class_name)
-        content = content.replace("custommanager", variable_name)
-        custom_file.write_text(content)
+        if custom_file.exists():
+            print(f"Error: {custom_file.name} already exists in this location.")
+        else:
+            variable_name = "_".join(word for word in command.lower().split())
+            class_name = "".join(word.capitalize() for word in command.split())
+            content = (Pathier(__file__).parent / "custom_manager.py").read_text()
+            content = content.replace("CustomManager", class_name)
+            content = content.replace("custommanager", variable_name)
+            custom_file.write_text(content)
 
     def _choose_db(self, options: list[Pathier]) -> Pathier:
         """Prompt the user to select from a list of files."""
